xsetwacom: allow removing special button mapping
If a user did a special button mapping, such as key-to-button,
and then changed it back to simple button-to-button mapping,
the special mapping still occured.

The main fix is to memset keys[] to clear out old mapping.

The other fix is more cosmetic.  We were leaving around a
property that was unreferenced by anything.  If a user
did an "xinput list-props pad", they way scratch there
head by old button mapping properties were still around.

Signed-off-by: Chris Bagwell <chris@cnpbagwell.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>