Init relative X/Y axis to use tablets min/max size
wcmRotateAndScaleCoordinates() uses these values to scale
values based on axis initialization.  The change to use -1..-1
caused it to scale in a way that cursor was overly accelerated.

This is a modified revert to a commit.  Before -1..-1
change, the code would set range to 0..maxX/Y but this
commit changes to match absolute mode behavior.  They both come
to same answer anyways.

Signed-off-by: Chris Bagwell <chris@cnpbagwell.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>