compute factorX/factorY from bottomX/bottomY.
maxWidth and maxHeight are no longer set and this causes
relative devices to not work (factor* = 0 which scales to
zero values always).

bottomX/bottomY is meant to be a value that optionally contains scaling
information that can be passed to xf86AxisScale() to scale
from tablet to screen ratios.  See wcmInitArea() for how that is done.

Derive the scaling factor from bottomX/bottomY.  Since this normally
1:1, the code is rarely used.

This patch fixes a regression where Bamboo touch device would not
move cursor any more.

Signed-off-by: Chris Bagwell <chris@cnpbagwell.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>