Add a testing framework for the driver.
This uses glib's testing framework provided glib-devel was available at
configure time (same as the X server's testing framework).

How this works:
The wacom driver is recompiled for the testing framework and linked into the
wacom-tests binary (there's a warning about that but ignore that). Since the
.so uses a number of symbols from the Xorg binary, these are provided as
stubs in fake-symbols.c. Additional binaries can be added through the
check_PROGRAMS automake variable.

The driver will be rebuild without static symbols for the test, allowing
each function call to be tested.

To run the tests, run 'make check'.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Acked-by: Ping Cheng <pinglinux@gmail.com>