Streamline the touch resolution
Resolution uses different units in the kernel (point/mm), XInput
(point/m), and the hardware protocol (point/cm/1000 for touch and
point/inch for pen).

We have a resolution table for pen in the X driver. But we get the
touch resolution directly from the kernel through HID descriptor.

For kernels older than 2.6.30, resolution is not part of absinfo.
For kernels older than 2.6.35, kernel does not pass resolution to
the userland.

This patch addresses those issues and correct a conversion bug
introduced by the RX/RY workaround.

Signed-off-by: Ping Cheng <pinglinux@gmail.com>
Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>