Prepare for Tablet PC MT format
We borrowed _TAP_ for 2FGT Tablet PC support when MT protocol was
not available (before kernel 2.6.30 was released). MT support in
the kernel is getting stable now. To move Tablet PC to MT format,
we face two issues:

1. USB Tablet PC 1FGT devices will not emit BTN_TOOL_FINGER event
since it is a touchscreen device.

2. Serial Tablet PC 2FGT devices do not emit BTN_TOOL_DOUBLETAP
event since they are not touchpads.

This patch add those keys back to the supported key array so we
can retrieve the proper keys when we need to decide the features
that are associated with the device.

Changes to v1:

1. Replace usbGenericTPCQuirks with usbGenericTouchscreenQuirks
2. Remove USB from the second item above

as suggested by Chris Bagwell.

Signed-off-by: Ping Cheng <pinglinux@gmail.com>
Acked-by: Chris Bagwell <chris@cnpbagwell.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>