Check if the device has been initalised before processing the events.
This is a stopgap solution only. Due to the current hotplugging approach,
the dependent tools are initialised before the parent device. This leaves us
with a time window where the parent device does not yet have a DeviceIntPtr
attached.

If an event from the parent device comes in during this time window, posting
the event causes segfaults. Fix this for now by checking if there is a
device attached to it yet.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>