Make 1 finger tap work based on finger release.
Previous behavior required a double-tap to get a single finger
tap to work and didn't always work for me.  Its behavior was basically
tap-and-drag gesture instead of tap-to-click.

Converted to standard touchpad behavior of single tap based on finger
release.  This is arguely the more important gesture to users.

Removed wcmTouchpadMode since it wasn't needed for simple tap-to-click
gesture.  If tap-and-drag gets added back and needs a variable, it can
be added back with a name that has "drag" in it to understand its
intended usage.

Signed-off-by: Chris Bagwell <chris@cnpbagwell.com>
Tested-by: Andrzej Giniewicz <gginiu@gmail.com>
Acked-by: Ping Cheng <pinglinux@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>