Don't discard the first 2 USB events
It was an issue with older kernels. This driver only supports
kernels 2.6.26 and later. No need to do that any more.

There is a special case that this code may have helped though.
If tool is on the tablet while driver started, waiting for more
events before processing them would fill some otherwise filtered
events from the kernel. Even for this case, the solution should
be added inside wcmUSB.c, instead of here.

We'll work on that in wcmUSB.c later. Just for the record. The
perfect solution would be to retrieve all events from the kernel
if important events, such as tool type, serial number, etc, are
missing. This would require quite some changes in wcmUSB.c. And,
the challenge is that linux-input doesn't provide the means for
us to retrieve all of the events that posted from the kernel. At
least, we can not retrieve serial number and MT events yet.

Signed-off-by: Ping Cheng <pinglinux@gmail.com>
Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>