Switch to a WorkProc for device hotplugging.
The current approach to hotplugging is to call NIDR directly from the parent
device's PreInit. This causes a nested device layout as the dependent
devices are initialised fully before the parent device is initialised.

Switch to a WorkProc instead. Copy the necessary information to create a new
device later and tell the server to call the actual hotplug function when it
has time.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Acked-by: Ping Cheng <pinglinux@gmail.com>