Keep an enabled flag in the tool to avoid racy startup.
There is a small time window where a device may try to send an event even
though it is not fully setup to send events yet, causing a server crash.

This window opens when the tool is added to the list of devices in
wcmParseOptions() and closes with the server calling xf86ActivateDevice().
If an event for a dependent device is processed during that time, the tool
will be available but the device pointer is still invalid.

Crash can be reproduced by putting a breakpoint after wcmParseOptions() for
the eraser, then generating events with the eraser. These will cause the
tool to dereference tool->device->dev, which is uninitialized.

Work around this with a simple "enabled" flag that is set whenever the tool
is actually enabled.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Acked-by: Ping Cheng <pinglinux@gmail.com>