Re-vamp source inclusion for test build suite.
test/Makefile.am included src/Makefile.am for the driver sources. However,
that also lead to the driver being built twice, once in src/ and once in
test/. On make install, test/wacom_drv.so would overwrite the
src/wacom_drv.so and thus install a driver with all static symbols compiled
as non-static.

This caused the VCOPY bug worked around in
9d227cd0bd985e7dcb7923974361145ba1d78713.

Avoid this by moving the sources into a shared included Makefile and build
an independent libwacom-test.la in the test directory.

Reported-by: Eduard Hasenleithner <ehasenle@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Tested-by: Eduard Hasenleithner <ehasenle@gmail.com>
Acked-by: Ping Cheng <pinglinux@gmail.com>