test: write a non-test for wcmTilt2R
It's non-transparent what wcmTilt2R actually does at the moment, so generate
a test table from the current code and test against that. This
way we at least know when we're breaking something.

Exception: wraparound value generates +900 instead of -900.  We must not
send +900, our max range is 899, then we wrap again. This function isn't
supposed to generate this value, so manually fixed up.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Acked-by: Ping Cheng <pinglinux@gmail.com>