Remove one FIXME
Get the "f" out of the code as well as the root cause.

A proper fix would be to use resolution to decide the distance.
That can be considered when we push the resolution table to
the kernel. Hopefully, at that time, we can also post MT events
to XInput 2.1 and push the gesture support to the desktop.

Signed-off-by: Ping Cheng <pinglinux@gmail.com>
Acked-by: Chris Bagwell <chris@cnpbagwell.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>