Enable left click for MT touchscreen
Tools that are in absolute mode emit left click when the tool touches
the surface. However, tools in relative mode need to locate the object
on the surface before it selects (left click on) the object. That is,
a second event (tap/leave/touch) is needed to trigger a left click for
relative mode tools.

Tools on touchscreen is in absolute mode. So, send left click
down/up when the first finger touches/leaves the tablet.

Signed-off-by: Ping Cheng <pinglinux@gmail.com>
Acked-by: Chris Bagwell <chris@cnpbagwell.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>