Prepare for the removal of BTN_TOOL_FINGER for pad tool in the kernel
BTN_TOOL_FINGER was borrowed to group PAD events. With the support
of pen and touch on the same logical port, we can not use it any more.

PAD events of generic protocol has already manually routed to the pad
channel, the last channel of wcmChanel.

This patch groups PAD events for protocol 4 and 5 by the device_id
sent through ABS_MISC event. In/out proximity event is also set by
the ABS_MISC event.

Signed-off-by: Ping Cheng <pinglinux@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>