Move findDeviceType from wcmCommon to wcmUSB
Serial model is not supported. So, findDeviceType is only used by
wcmUSB.c.

Now it is the individual model's, wcmUSB or wcmISDV4, responsibility
to set the device type locally. For wcmUSB, since kernel input filters
out duplicated events, we should retrieve the device type and
associated events directly from the kernel to cover the case when tool
is on the tablet during driver starts.

wcmCommon.c will treat packet with undefined device_type as invalid.

Signed-off-by: Ping Cheng <pinglinux@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>