Resolve the oldHwProx puzzle
Change oldHwProx to oldCursorHwProx to better reflect its use.

oldCursorHwProx keeps the hardware in/out prox state so we can
set the MaxCursorDist for the next round of relative cursor
movement when tool first comes in prox.

For Protocol 5 devices, distance is reported as the maximum
when tool is first detected. Then it is gradually reduced to 0
when getting closer to the tablet. Protocol 4 is the opposite:
distance is 0 when tool is first in prox. The value increases
as the tool gets closer to the tablet.

The maximum distance (wcmMaxDist) is model specific. They are
retrieved from the kernel.

Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Ping Cheng <pinglinux@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>