Allow parse_actions to parse simple button maps
If argv contains only a single numeric argument, we assume the
intent is to perform a button press. We add the necessary keywords
to make the parser happy, and then parse as normal.

This allows map_wheels to handle the "raw button" case that it has
been missing. While map_button_simple also does the same* thing,
that's another commit :)

*Actually, map_button_simple does /almost/ the same thing, but I
highly doubt anyone relies on the subtle difference in behavior.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>