Merge 'map_wheels' and 'map_button' into a unified 'map_actions'
These two functions are very similar, and have been merged together
to provide a single conduit for all action mappings. Button-handling
code requires a little bit of special-casing (we need to be sure to
get a button to modify, and to change the arguments appropriately),
but nothing remotely worthy of breaking into its own function.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>