Revamp button/wheel/strip getters
'get_map' is the new getter entrypoint and attempts to determine
the mapping present on buttons, wheels, and strips. It calls on
'get_actions' (essentially get_special_button_map) and 'get_button'
(an 8-bit XA_INTEGER version of the same) to determine this. If
both fail to find a mapping, it falls back to the method originally
used for non-actions: punt and ask XGetDeviceButtonMapping for
something.

When raw buttons are removed from the driver, it should be possible
to replace 'get_map' with 'get_actions' (since everything will be
an action at that point).

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>