Have wheel and strip properties be "actions" instead of "buttons"
Continuing with removing raw button code, we create the wheel and
strip properties ready to hold actions. Prior to this, the properties
would be initialized with the raw buttons; before xsetwacom could
set them to actions they would need to be converted.

This does away with the conversion step by following the mold already
in place for buttons. With no more need for conversion, we can also
nuke the conversion-related bits from xsetwacom.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>