Add ToolSerials option for hotplugging pens
Provide a new "ToolSerials" option which describes tools that should
be automatically hotplugged along with the standard devices. This is
achieved by parsing "ToolSerials" while the master device is being
initialized to create a WacomTool list. This list is then used while
hotplugging child devices to create tool-specific devices with the
"Serial" option set appropriately.

The "ToolSerials" option should be added to the 50-wacom.conf file
in /etc/X11/xorg.conf.d so that it is applied to hotplugged tablets.
The option takes the following format:

Option "ToolSerials" "serial[,type[,label]][;...]"

Or, for a more concrete example:

Option "ToolSerials" "113246231,airbrush,Airbrush;85983311,artpen,Artpen"

ToolSerials recognizes the "pen", "airbrush", "artpen", and "cursor"
types. Unknown (or unspecified) types default to creating "pen"-type
tools with both 'stylus' and 'eraser' devices.

Signed-off-by: Alexia Death <alexiadeath@gmail.com>
Acked-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>

Squashed in:
Have ToolSerials patch re-use existing infrastructure

This tweaks the prior ToolSerials patch to use the existing
infrastructure where appropriate. For instance, using the 'WacomTool'
struct instead of the custom 'WacomToolSerial' struct.

This should actually be integrated as a "fixup", but is included
here to more easily see the changes from the original.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>