Get the latest values for serial numbers when a client asks for them.
Whenever a client requests the property values, update them from the current
device state. Due to race conditions with the SIGIO handler, we have a
window between updating the device state (during the handler) and updating
the property and sending events (requires alloc, thus not in the handler).

If a client asks for the property value, update from the state so the window
only affects property events, not actual values.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>