xsetwacom: Support MapToOutput for TwinView.
The NVIDIA binary driver doesn't support RandR 1.2 but it does support
Xinerama. Due to a server bug, we still get the current server's RandR
version back when we query, so we need to check to the binary
driver-specific NV-CONTROL extension first. If that or a RandR < 1.2
version is present fall back to Xinerama if we need to.

User commands:
    xsetwacom set "device name" MapToOutput HEAD-0
    xsetwacom set "device name" MapToOutput HEAD-1

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Tested-by: Jason Gerecke <killertofu@gmail.com>