Default to touch tool for multi-touch devices
We really need to add hooks to query pre-existing device state when
input driver starts up because slot ID is almost never sent when
touching 1 finger to screen.

In mean time, it makes sense to default device_type to most likely
value of TOUCH (the T in MT) for this function used only by MT
devices.

The default of zero is causing some events to be filtered mistakenly.

Signed-off-by: Chris Bagwell <chris@cnpbagwell.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>