Don't adjust the tablet area property for multimonitor.
The interactions between tablet_area in multimonitor setups are
complicated to say the least. Amongst other issues, it caused the
calibration to be out by a bit and rather unpredictable.

Force the area property to always be in device coordinates. This means
that clients that need this property now also need to set the property
in device coordinates which may be difficult if the axis ranges have
already been adjusted.

Only way for clients: reset the property with -1 values, read the
values, then set from there.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>