Allow 'MapToOutput' to accept X geometry strings
This patch allows the MapToOutput command to accept fully-
specified X geometry strings as valid output names. The
XParseGeometry function describes these strings as having
the folowing format:

[=][<width>{xX}<height>][{+-}<xoffset>{+-}<yoffset>]

Strings with width, height, xoffset, and yoffset defined
will be accepted and the device mapped to the rectangle
it describes.

This patch also renames the function '_set_matrix' to bring
its naming in line with the other set_output_XXXX functions.