Make InitWcmAtom support arbitrary property types
InitWcmAtom could only create properties of type XA_INTEGER or
XA_ATOM. This commit adds support for any atom type by means
of adding a type parameter to InitWcmAtom. All instances of
InitWcmAtom use are adapted to the new parameters.

This is a preparation for a new parameter of type XA_PIXMAP
needed for the upcoming OLED commit.

Signed-off-by: Eduard Hasenleithner <eduard@hasenleithner.at>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>