Add 0x10804 to recognized Art Pen device IDs
Intuos4 Art Pens may have an ID of 0x10804 for their tip and
0x1080C for their eraser. There may be other IDs which need
matching as well, but that is a "to investigate" item for the
moment.

Note that due to a mostly-benign issue in the kernel driver, the
highest nybble of the ID is placed one nybble *too* high. Until
a proper fix is made and tested for the kernel, we need to check
for an ID of "0x100804" instead of "0x10804".

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>