Use xf86PostButtonEventP for held button action releases
It appears that commit 485137ef partially missed this particular
call in its mass-migration away from xf86PostButtonEvent. This
results in the pointer jumping away from the correct location
when a held button action is finally released. It is most noticible
when the action "button +1" is assigned to the stylus or eraser and
a window is dragged.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>