Revert "Only convert I4 tilt bits to rotation for the I4 mouse"
This reverts commit 9be9719d78af035c45ea4b191c21fab0ee54f198.
Under the right conditions, the 'tool' pointer can be NULL,
leading to the X server crashing. An easy way to trigger the
crash is to spam pad events (such as button presses) while
holding the stylus on the edge of proximity.

Rather than fix this with additional checks for NULL, Ping
clued me into a much more elegant (and bug-free) way to
implement the check for an I4 mouse. See next patch.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>