Migrate Tablet(Has|Set)Feature to use the wacom-util macros
In addition to consolidating code, this happens to fix a bug
in the implementation of TabletHasFeature which would return
true provided *any* bit of a multi-bit mask was set. Ever
since 7c667678 when WCM_TPC was changed to a multi-bit mask,
tablets with WCM_LCD set have been (mis-)treated as tablet PCs.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>