Have usbInitToolType return the old tool type if available
If the tool was not specified within an event packet, the
usbInitToolType function would return TOUCH_ID. This caused
problems for combined pen+touch devices when the pen was in
range, since only the first packet (which explicitely specified
the STYLUS tool) would not be filtered as "single touch"
data.

This patch passes along the old tool type, so that if no
tool is found in the current packet, the old tool type may
be adopted.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Reviewed-by: Chris Bagwell <chris@cnpbagwell.com>