Don't share names of action atoms across properties
The names of action atoms are currently created without regard to
the property they will be stored in. This results in both e.g.
button 1 and StripLeftUp using atoms with the name "Wacom button
action 1". When setting either button 1 or StripLeftUp, we ask
the server for an atom named "Wacom button action 1" and in both
cases receive the *same* atom. This means button 1 is nothing
more than an alias for StripLeft up (and similarly for the other
actions which share the same name).

This patch uses the name of the property the action will be stored
in as the prefix to prevent this aliasing from occuring.