Reset-on-None for Actions property
This patch adds the ability to reset an individual Action by
setting the corresponding element of an Actions property to
the 'None' atom (i.e. 0). When encountered, the None will be
replaced with an atom representing the default Action. This
will let the driver behave as it currently does once raw
button support is removed.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>