Find mouse buttons on pad devices if no generic buttons found
Devices like the Graphire4 and Bamboo-series have Pad buttons
which emit "mouse button" events (e.g. BTN_FORWARD) instead of
the usual "generic button" events (e.g. BTN_0). Currently, the
code considers the former to be valid for the Pad device only
if the tablet doesn't also support a Cursor device. Devices
with mouse buttons on the Pad *and* have a Cursor device
(like the Graphire4 and very early Bamboos) end up declaring
zero Pad buttons because of this, causing us to allocate
an insufficient number of actions and thereby prevent the
re-mapping of their forward and back buttons.

This patch changes the condition from "no Cursor device exists"
to "no generic buttons exist". In theory this can result in the
Pad claiming it has buttons when none physically exist, but it
shouldn't result in any problems.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>