Add support for Cintiq 24HD touch
Adds support for the EMR and touch digitizers on the Cintiq 24HD
touch. The digitizers are on separate USB devices, which can lead to
the cursor jumping back and forth between the stylus and finger.
Performing arbitration between the two (disabling finger while pen is in
proximity) is expected to be handled by the kernel driver.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>