Have keysym_to_keycode handle unused keycodes
The current implementation of keysym_to_keycode does not
properly handle unused keycodes. The functions XkbKeySymsPtr
and XkbKeyGroupWidth mislead it into believing it is reading
a keysym associated with the unused keycode when in reality
it is reading a keysym associated with the *next used*
keycode.

This impacts different keys depending on the keyboard layout
in use. On mine for instance, the left Super key is unusable
because its keycode is preceeded by an unused keycode.

To resolve this issue, we take a page from XKeysymToKeycode
and use the function XKeycodeToKeysym instead of manually
accessing the keysyms.

https://sourceforge.net/tracker/?func=detail&aid=3555424&group_id=69596&atid=525124

Reported-by: alexmaret
Signed-off-by: Jason Gerecke <killertofu@gmail.com>