Ignore out-of-prox events from unknown devices already out-of-prox
Fixes an issue with the Intuos5, which sends extra out-of-prox
notifications when the ExpressKeys are touched. This causes the driver
to ask the kernel for the tool currently in prox and post an empty event
for it. The result is a cursor that (only while the pen is in prox)
sometimes jumps to the top left when touching the buttons.

We fix this by ignoring events that claim to be leaving prox when the
tool is already supposedly out of prox.

https://bugs.freedesktop.org/show_bug.cgi?id=54250

Reported-by: Timo Aaltonen <tjaalton@ubuntu.com>
Signed-off-by: Jason Gerecke <killertofu@gmail.com>