Fix double-free in wcmPreInitParseOptions
The 'tool' function here is set equal to 'priv->tool' earlier in
this function. If an error condition occurs, the tool can be freed
both here and in wcmFree (via wcmUninit).

It should be safe to postpone this free() until wcmFree to be
freed along with the rest of the memory allocated in wcmAllocate.
There should be no leak (wcmUninit will be called on failure) and
any code relying on priv->tool being a valid pointer after the
failure of wcmPreInitParseOptions is already broken.

https://bugs.freedesktop.org/show_bug.cgi?id=55200

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>