Fix prerequisite check in wcmUpdateSerial
The prerequisite check in wcmUpdateSerial is incorrect. If the tool ID
changes, the property reflecting it will not be updated unless/until the
serial number also changes. The check is updated to reflect this.

Additionally, the prerequesite check at one of its callsites is removed
since the function performs the same check internally.

https://bugs.freedesktop.org/show_bug.cgi?id=45237

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Reviewed-by: Ping Cheng <pingc@wacom.com>