Increase maximum allowed X11 button number
Actions specify X11 button numbers that should be generated when
performed. The maximum button number that could be used in an event
has been raised from the driver-imposed limit (WCM_MAX_BUTTON) to
the server-imposed limit (WCM_MAX_X11BUTTON).

This feature is necessary for gnome-settings-daemon to work properly.
To reliably identify each touch strip and ring, g-s-d assigns each
a high button numbers to send for each direction. Without this
patch, the driver ignores the actions set by g-s-d and continues
to send whatever the default action is instead.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>