Support pen/touch arbitration across product ids
Devices with different product ids for pen and touch, such as
Cintiq 24HD and DTH 2242, were added recently. Applying arbitration
only for tools with same product id is not enough.

However, we want to make sure tools with same product id are lined
properly first. That's why the same routine is repeated instead of
combined.

Signed-off-by: Ping Cheng <pingc@wacom.com>
Reviewed-by: Jason Gerecke <killertofu@gmail.com>