strdup the option key/values in our input_option_new
This is a interface for servers with ABI < 14 (up to server 1.11).
The server version of it strdups the values, so we must do the same.

Otherwise the values assigned to the InputOption list in
wcmOptionDupConvert are freed during xf86OptionListFree().

That later causes a SIGABORT when NewInputDeviceRequest starts using
those options.

https://sourceforge.net/p/linuxwacom/bugs/233/

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Acked-by: Ping Cheng <pinglinux@gmail.com>