Correct return value of special_map_button
The 'special_map_*' functions called by 'parse_actions' are expected
to return a count of the number of words of input they have consumed.
The 'special_map_buttom' function errounously returns the number of
actions added instead, causing the program to potentially skip over
requested actions which follow a list of "button" actions.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>