HACK: Define 'input_option_new' for Ubuntu 12.04 'frankenserver'
The 'frankenserver' X server that ships with Ubuntu 12.04 doesn't
actually implement the input_option_new function, despite what its
version number would imply. With the symbol missing, the server
crashes once a tablet is plugged in.

This hack forces the input_option_new function to be compiled,
allowing it to be resolved on the frakenserver. This patch DOES NOT
belong in the upstream xf86-input-wacom tree.