conf: add N-Trig DuoSense to our matched devices
That device has a pen and an eraser tool, as well as a touch device. Handle it
with the wacom driver by default, the evdev driver isn't quite up to scratch
here.

This also fixes a missing match in the .fdi file for the base N-Trig devices -
just in case.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>