Improve wcmLinkTouchAndPen device matching
If a tablet has pen and touch interfaces, they appear to the system as
seperate devices. Because it is necessary for these devices to share
information, `wcmLinkTouchAndPen` was introduced to connect the two
"halves" of the single physical tablet together (similar to
how `wcmMatchDevice` links logical devices together).

`wcmLinkTouchAndPen` has a few bugs in its implementation which prevent
it from doing its job as well as it should. This patch introduces a new
funcation named `wcmIsSiblingDevice` which is responsible for
determining if two devices are part of the same tablet. It includes a
slightly stricter (but still imperfect) check to reduce the number of
false-positives in the linking process.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>