Log when the pen device appears to be too worn out
Worn out pens have an initial pressure != 0. If this pressure
persists while the pen is in proximity and still exists when
going out of proximity, warn about worn out tool.
Some devices seem to send a zero pressure when going out of
proximity. Thus we record the last pressure when still in
proximity.
Also we count the number of events we have received while in
proximity so the risk of false reports on 'quick taps' is lower.

This condition may also occur intermittently on abusive operation.

Signed-off-by: Egbert Eich <eich@suse.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>