Make tested static functions non-static for test suite
This gets rid of the hack which redefines 'static' to make
every static function in the driver non-static for the test
suite.
Instead we use a macro instead of 'static' in the function definition
of each function used in the test suite which we can set to empty
when compiling the test suite.
At the same time we create a declaration for each of those functions
which gets us rind of about a bazillion compiler warnings.

Signed-off-by: Egbert Eich <eich@suse.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>