Get rid of spurious 'format not a string literal' warnings
With '-Wformat-nonliteral' gcc kindly warns us that it cannot check the
arguments of a scanf statement as the format string is not literal.
This happens if the format string itself is pointer.
Since the list of format strings only has two elements the same thing
can be achieved in code directly.

Signed-off-by: Egbert Eich <eich@suse.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Jason Gerecke <killertofu@gmail.com>